requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.2
appdirs==1.4.4
apprise==1.2.0
argon2-cffi==20.1.0
asgi-lifespan==2.0.0
asgiref==3.5.2
asttokens==2.1.0
astunparse==1.6.3
async-exit-stack==1.0.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bentoctl==0.3.4
bentoml==1.0.8
bleach==5.0.1
blinker==1.4
boto3==1.26.3
botocore==1.29.3
Brotli==1.0.9
brotlipy==0.7.0
build==0.9.0
cached-property==1.5.2
cachetools==4.2.2
cattrs==22.2.0
Cerberus==1.3.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
circus==0.17.1
click==8.0.4
cloudpickle==2.2.0
colorama==0.4.6
commonmark==0.9.1
ConfigArgParse==1.5.3
contextlib2==21.6.0
coolname==2.0.0
cramjam==2.6.2
croniter==1.3.8
cryptography==38.0.1
cycler==0.11.0
databricks-cli==0.17.3
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
deepdiff==6.2.1
deepmerge==1.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
docker==6.0.0
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.2.0
fastapi==0.87.0
fastjsonschema==2.16.2
fastparquet==0.8.3
filelock==3.8.0
Flask==2.1.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.2.0
fs==2.4.16
fsspec==2022.11.0
future==0.18.2
gast==0.5.3
gevent==22.10.2
geventhttpclient==2.0.8
gitdb==4.0.9
GitPython==3.1.29
google-auth==1.21.3
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==2.0.0.post0
griffe==0.24.1
grpcio==1.42.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
httpcore==0.16.1
httpx==0.23.1
hyperframe==6.0.1
hyperopt==0.2.7
idna==3.4
importlib-metadata==4.11.3
importlib-resources==5.2.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.0
jupyter==1.0.0
jupyter_client==7.4.4
jupyter-console==6.4.3
jupyter_core==4.11.2
jupyter-server==1.21.0
jupyterlab==3.3.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.2
jupyterlab-widgets==1.0.0
keras==2.8.0
keras-image-helper==0.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kubernetes==25.3.0
libclang==14.0.6
locust==2.13.0
Mako==1.2.4
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mlflow==1.30.0
msgpack==1.0.4
multidict==6.0.2
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
oauthlib==3.2.1
opentelemetry-api==1.13.0
opentelemetry-instrumentation==0.34b0
opentelemetry-instrumentation-aiohttp-client==0.34b0
opentelemetry-instrumentation-asgi==0.34b0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
opentelemetry-util-http==0.34b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.2
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pathspec==0.10.2
pendulum==2.1.2
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pip-requirements-parser==31.2.0
pip-tools==6.9.0
pipenv==2022.11.25
platformdirs==2.5.4
ply==3.11
prefect==2.6.9
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.4.0
PyNaCl==1.5.0
pynvml==11.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-multipart==0.0.5
python-slugify==7.0.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==4.6.0
querystring-parser==1.2.4
readchar==4.0.3
requests==2.28.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==12.6.0
roundrobin==0.0.4
rsa==4.7.2
s3transfer==0.6.0
schema==0.7.5
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
semantic-version==2.10.0
Send2Trash==1.8.0
setuptools==65.4.0
shellingham==1.5.0
simple-di==0.1.5
simple-term-menu==0.4.4
sip==6.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.6.0
starlette==0.21.0
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-serving-api==2.7.0
termcolor==1.1.0
terminado==0.17.0
testpath==0.6.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typer==0.7.0
typing_extensions==4.3.0
Unidecode==1.3.6
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
watchdog==2.1.6
watchfiles==0.18.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.14.1
xgboost==1.7.1
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.5.1
